# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "riot-proxy",
    "version": "1.0.0",
...
    "license": {
      "name": "MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A consumer key, `rpx_…`, issued by `POST /v1/admin/consumers` and shown exactly once. Only its SHA-256 is stored."
      },
      "tokenQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "**WebSocket handshake only.** A browser `WebSocket` cannot set an `Authorization` header, so `/v1/ws` takes the same key as `?token=`. Do not use this on the HTTP routes — a key in a query string ends up in access logs."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "QUOTA_EXCEEDED",
                  "NOT_FOUND",
                  "UPSTREAM_ERROR",
                  "RATE_LIMITED",
                  "BAD_REGION",
                  "VALIDATION",
                  "INTERNAL"
                ]
              },
              "message": {
                "type": "string"
              },
              "retryAfter": {
                "type": "number"
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/admin/consumers": {
      "post": {
        "tags": [
          "admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 100,
                    "type": "string"
                  },
                  "scopes": {
                    "minItems": 1,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "read",
                        "admin"
                      ]
                    }
                  },
                  "quotaPerMin": {
                    "minimum": 1,
                    "maximum": 1000000,
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Unknown platform or region": {
                    "summary": "Unknown platform or region",
                    "value": {
                      "error": {
                        "code": "BAD_REGION",
                        "message": "params/platform must be equal to one of the allowed values"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Missing or unrecognised key": {
                    "summary": "Missing or unrecognised key",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "message": "Missing or invalid consumer key"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Key lacks the scope (or the admin IP allowlist rejected it)": {
                    "summary": "Key lacks the scope (or the admin IP allowlist rejected it)",
                    "value": {
                      "error": {
                        "code": "FORBIDDEN",
                        "message": "This key does not carry the admin scope"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "No such resource": {
                    "summary": "No such resource",
                    "value": {
                      "error": {
                        "code": "NOT_FOUND",
                        "message": "No account for that Riot ID"
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Your quota — slow down. Sent with Retry-After": {
                    "summary": "Your quota — slow down. Sent with Retry-After",
                    "value": {
                      "error": {
                        "code": "QUOTA_EXCEEDED",
                        "message": "Quota of 600/min exceeded",
                        "retryAfter": 37
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-badges": [
          {
            "name": "Admin scope"
          }
        ]
      },
      "get": {
        "summary": "List consumers",
        "tags": [
          "admin"
        ],
        "description": "Every consumer, including revoked ones — `disabledAt` distinguishes them. The key itself is never returned: only its sha256 is stored, and not even that is exposed here. A plaintext key is shown exactly once, by `POST /v1/admin/consumers`.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "consumers"
                  ],
                  "properties": {
                    "consumers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "name",
                          "scopes",
                          "quotaPerMin",
                          "createdAt",
                          "disabledAt"
                        ],
                        "properties": {
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "scopes": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "enum": [
                                "read",
                                "admin"
                              ]
                            }
                          },
                          "quotaPerMin": {
                            "description": "Requests per minute allowed to this consumer",
                            "type": "integer"
                          },
                          "createdAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "disabledAt": {
                            "anyOf": [
                              {
                                "format": "date-time",
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ],
                            "description": "Set when the key was revoked; revoked consumers are kept, never deleted"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Unknown platform or region": {
                    "summary": "Unknown platform or region",
                    "value": {
                      "error": {
                        "code": "BAD_REGION",
                        "message": "params/platform must be equal to one of the allowed values"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Missing or unrecognised key": {
                    "summary": "Missing or unrecognised key",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "message": "Missing or invalid consumer key"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Key lacks the scope (or the admin IP allowlist rejected it)": {
                    "summary": "Key lacks the scope (or the admin IP allowlist rejected it)",
                    "value": {
                      "error": {
                        "code": "FORBIDDEN",
                        "message": "This key does not carry the admin scope"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "No such resource": {
                    "summary": "No such resource",
                    "value": {
                      "error": {
                        "code": "NOT_FOUND",
                        "message": "No account for that Riot ID"
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Your quota — slow down. Sent with Retry-After": {
                    "summary": "Your quota — slow down. Sent with Retry-After",
                    "value": {
                      "error": {
                        "code": "QUOTA_EXCEEDED",
                        "message": "Quota of 600/min exceeded",
                        "retryAfter": 37
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-badges": [
          {
            "name": "Admin scope"
          }
        ]
      }
    },
    "/v1/admin/consumers/{id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Unknown platform or region": {
                    "summary": "Unknown platform or region",
                    "value": {
                      "error": {
                        "code": "BAD_REGION",
                        "message": "params/platform must be equal to one of the allowed values"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Missing or unrecognised key": {
                    "summary": "Missing or unrecognised key",
                    "value": {
                      "error": {
                        "code": "UNAUTHORIZED",
                        "message": "Missing or invalid consumer key"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Key lacks the scope (or the admin IP allowlist rejected it)": {
                    "summary": "Key lacks the scope (or the admin IP allowlist rejected it)",
                    "value": {
                      "error": {
                        "code": "FORBIDDEN",
                        "message": "This key does not carry the admin scope"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "No such resource": {
                    "summary": "No such resource",
                    "value": {
                      "error": {
                        "code": "NOT_FOUND",
                        "message": "No account for that Riot ID"
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Your quota — slow down. Sent with Retry-After": {
                    "summary": "Your quota — slow down. Sent with Retry-After",
                    "value": {
                      "error": {
                        "code": "QUOTA_EXCEEDED",
                        "message": "Quota of 600/min exceeded",
                        "retryAfter": 37
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "x-badges": [
          {
            "name": "Admin scope"
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    },
    {
      "url": "{scheme}://{host}",
      "description": "Your deployment",
      "variables": {
        "scheme": {
          "enum": [
            "https",
            "http"
          ],
          "default": "https"
        },
        "host": {
          "default": "riot-proxy.example.com"
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}